% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcqmcint.R
\name{digitalnet.points}
\alias{digitalnet.points}
\title{get points from Digital Net}
\usage{
digitalnet.points(digitalNetID, dimR, dimF2 = 10, count,
  digitalShift = FALSE)
}
\arguments{
\item{digitalNetID}{1:Niederreiter-Xing low WAFOM, 2:Sobol low wafom,
3:Sobol large dimension.}

\item{dimR}{dimention.}

\item{dimF2}{F2-dimention of each element.}

\item{count}{number of points.}

\item{digitalShift}{use digital shift or not.}
}
\value{
matrix of points where every row contains dimR dimensional point.
}
\description{
DigitalNetID:
\itemize{
\item{1:}{Niederreiter-Xing low WAFOM}
\item{2:}{Sobol low WAFOM}
\item{3:}{Sobol Point Set up to dimension 21201}
}
}
