% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcqmcint.R
\name{mcint}
\alias{mcint}
\title{Monte-Carlo Integration}
\usage{
mcint(integrand, N, s, m = 10, probability = 0.99)
}
\arguments{
\item{integrand}{integrand function.}

\item{N}{number of repeat.}

\item{s}{dimention, s should be 4 <= s <= 10.}

\item{m}{use 2^m samples.}

\item{probability, }{should be one of 0.95, 0.99, 0.999, or 0.9999.}
}
\value{
integrated mean value and absolute error.
}
\description{
Compute Monte-Carlo Integration.
}
\details{
integrand should receive numeric vector of length s and
should return numeric value.
}
