% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcqmcint.R
\docType{package}
\name{rmcqmcint-package}
\alias{rmcqmcint-package}
\alias{rmcqmcint}
\title{Monte-Carlo and Quasi Monte-Carlo Integration}
\description{
Compute Monte-Carlo integration and Quasi Monte-Carlo Integration with
Low WAFOM Digital Net.
}
