% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcqmcint.R
\name{qmcint}
\alias{qmcint}
\title{Quasi Monte-Carlo Integration with Low WAFOM Digital Net}
\usage{
qmcint(integrand, N, s, digitalNetID = 1, m = 10, probability = 0.99)
}
\arguments{
\item{integrand}{integrand function.}

\item{N}{number of repeat.}

\item{s}{dimention, s should be 4 <= s}

\item{digitalNetID}{1:Niederreiter-Xing low WAFOM, 2:Sobol low wafom,
3:Sobol large dimension.}

\item{m}{F2-dimention of each element, m should be 10 <= m <= 18.}

\item{probability, }{should be one of 0.95, 0.99, 0.999, or 0.9999.}
}
\value{
integrated mean value and absolute error.
}
\description{
Compute Quasi Monte-Carlo Integration with Low WAFOM Digital Net,
Niederreiter-Xing and Sobol.
}
\details{
DigitalNetID:
\itemize{
\item{1:}{Niederreiter-Xing low WAFOM up to dimension 10}
\item{2:}{Sobol low WAFOM up to dimentison 10}
\item{3:}{Sobol Point Set up to dimension 21201}
}

integrand should receive numeric vector of length s and
should return numeric value.
}
